Added SharpTest

+++ src/main/kotlin/com/thebluehats/server/game/enchants/Sharp.kt
@@ -3,7 +3,6 @@ package com.thebluehats.server.game.enchants
 import com.google.inject.Inject
 import com.thebluehats.server.game.managers.combat.CalculationMode
 import com.thebluehats.server.game.managers.combat.DamageManager
-import com.thebluehats.server.game.managers.combat.templates.DamageEnchantTrigger
 import com.thebluehats.server.game.managers.combat.templates.EnchantHolder
 import com.thebluehats.server.game.managers.combat.templates.PlayerDamageTrigger
 import com.thebluehats.server.game.managers.enchants.DamageTriggeredEnchant
@@ -11,7 +10,6 @@ import com.thebluehats.server.game.managers.enchants.EnchantGroup
 import com.thebluehats.server.game.managers.enchants.EnchantProperty
 import com.thebluehats.server.game.managers.enchants.processedevents.DamageEventEnchantData
 import com.thebluehats.server.game.utils.EnchantLoreParser
-import com.thebluehats.server.game.utils.EntityValidator
 import org.bukkit.Material
 import java.util.*
 
@@ -19,7 +17,7 @@ class Sharp @Inject constructor(private val damageManager: DamageManager, player
     DamageTriggeredEnchant(
         arrayOf(playerDamageTrigger), arrayOf(damageManager)
     ) {
-    private val percentDamageIncrease = EnchantProperty(0.04f, 0.07f, 0.12f)
+    private val percentDamageIncrease = EnchantProperty(0.04, 0.07, 0.12)
 
     override val name: String get() = "Sharp"
     override val enchantReferenceName: String get() = "Sharp"
@@ -34,7 +32,7 @@ class Sharp @Inject constructor(private val damageManager: DamageManager, player
         val level = data.level
         damageManager.addDamage(
             event,
-            percentDamageIncrease.getValueAtLevel(level).toDouble(),
+            percentDamageIncrease.getValueAtLevel(level),
             CalculationMode.ADDITIVE
         )
     }
